# api/openapi/customer.v1.yaml
openapi: 3.0.3
info:
  title: Customer API
  version: "1.0.0"
  description: >
    Delivery contract for the customer bounded context
    (internal/customer/interfaces/http).
servers:
  - url: http://localhost:8080
paths:
  /api/v1/customers:
    post:
      operationId: createCustomer
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCustomerResponse"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      operationId: listCustomers
      summary: List customers (from the read model)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerPage"
  /api/v1/customers/{id}:
    get:
      operationId: getCustomer
      summary: Get one customer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/customers/{id}/email:
    patch:
      operationId: changeCustomerEmail
      summary: Change a customer's email address
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204": { description: No Content }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    CreateCustomerRequest:
      type: object
      required: [name, email]
      properties:
        name: { type: string, maxLength: 200 }
        email: { type: string, format: email }
    CreateCustomerResponse:
      type: object
      properties:
        id: { type: string }
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        status: { type: string, enum: [active, deactivated] }
        created_at: { type: string, format: date-time }
    CustomerPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Customer" }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }